feat: Integrate Supabase SSR client and refactor client-related functions
