+++ package.json
@@ -13,6 +13,7 @@
     "@radix-ui/react-avatar": "^1.1.0",
     "@radix-ui/react-dialog": "^1.1.1",
     "@radix-ui/react-slot": "^1.1.0",
+    "@supabase/ssr": "latest",
     "@supabase/supabase-js": "^2.45.1",
     "@tabler/icons-react": "^3.12.0",
     "@vercel/analytics": "^1.4.1",
