+++ src/actions/coffees.ts
@@ -1,10 +1,12 @@
 "use server";
 import { TABLE_NAME } from "@/lib/constants";
-import { supabase } from "@/supabase.config";
+import { createClient } from "@/supabase-server";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
 export const addCoffee = async (formData: FormData) => {
+  const supabase = await createClient();
+
   const id = formData.get("id") as string;
   const coffees = parseInt(formData.get("coffees") as string);
   await supabase.from(TABLE_NAME).upsert({ id, coffees: coffees + 1 });
@@ -12,12 +14,16 @@ export const addCoffee = async (formData: FormData) => {
 };
 
 export const resetCoffee = async (FormData: FormData) => {
+  const supabase = await createClient();
+
   const id = FormData.get("id") as string;
   await supabase.from(TABLE_NAME).upsert({ id, coffees: 0 });
   revalidatePath("/admin");
 };
 
 export const deleteCoffee = async (FormData: FormData) => {
+  const supabase = await createClient();
+
   const id = FormData.get("id") as string;
   await supabase.from(TABLE_NAME).delete().eq("id", id);
   revalidatePath("/admin");
