+++ src/auth.config.ts
@@ -1,9 +1,9 @@
 import type { NextAuthConfig } from "next-auth";
 import Google from "next-auth/providers/google";
-import { supabase } from "./supabase.config";
 import { removeAccents } from "./lib/utils";
 import { TABLE_NAME } from "./lib/constants";
 import { track } from "@vercel/analytics/server";
+import { createClient } from "./supabase-server";
 
 const adminEmails = [
   "izquierdonelson@gmail.com",
@@ -48,6 +48,7 @@ export const authConfig = {
       const { email, name, image } = session.user;
       const normalizedName = removeAccents(name ?? "");
       await track("session", { email, role });
+      const supabase = await createClient();
 
       const { data, error } = await supabase
         .from(TABLE_NAME)
